+++ app/jobs/anonymise_deactivated_users_job.rb
@@ -1,0 +1,9 @@
+class AnonymiseDeactivatedUsersJob
+  include Sidekiq::Job
+
+  def perform
+    User.deactivated.where("deactivated_at < ?", 5.years.ago).each do |user|
+      AnonymiseUser.new(user:).call
+    end
+  end
+end